server/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.3
numpy==2.3.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pillow==11.0.0
psutil==7.1.0
python-decouple==3.8
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
